requirements.txt
coverage==3.7.1
Django==1.8.1
django-extensions==1.5.3
djangorestframework==3.1.1
gnureadline==6.3.3
ipython==3.1.0
mock==1.0.1
nose==1.3.6
pytz==2015.2
requests==2.7.0
scorched==0.7
six==1.9.0
